move unit tests to new folder
