+++ test-unit/bot/detectors/detector-test.utils.mjs
